Masked client ID and fixed game context detection
For some reason the "private" event on group.socket does not work for
me, so I've found a workaround with the "you" event, although didn't
bother checking if the group is private or not.
Fixed also the game detection using tagproConfig (window.port wasn't
working for me)

I'm using chrome to test this btw, don't know if it's a browser issue.